+++ odoo/songs/install/logistics.py
@@ -44,9 +44,13 @@ def activate_options(ctx):
 
 
 @anthem.log
-def set_delivery_pick_ship(ctx):
-    """ Setting pick-ship on the warehouse """
-    ctx.env.ref('stock.warehouse0').delivery_steps = 'pick_ship'
+def warehouse_settings(ctx):
+    wh_vlb = ctx.env.ref('stock.warehouse0')
+    wh_vlb.write({
+        'name': 'Villers-Le-Bouillet',
+        'code': 'VLB',
+        'delivery_steps': 'pick_ship',
+    })
 
 
 @anthem.log
@@ -244,6 +248,16 @@ def create_putaway(ctx):
                 '__setup__.stock_location_onorder').id,
         }
     )
+    create_or_update(
+        ctx, 'stock.fixed.putaway.route.strat',
+        '__setup__.stock_putaway_input_froid',
+        {
+            'putaway_id': ref('__setup__.stock_putaway_input').id,
+            'route_id': ref('__setup__.stock_location_route_pick_froid').id,
+            'fixed_location_id': ref(
+                '__setup__.stock_location_parking_frigo').id,
+        }
+    )
 
     parking_strat = [
         ('__setup__.stock_putaway_strat_parking_medoc',
@@ -254,10 +268,7 @@ def create_putaway(ctx):
          '__setup__.stock_location_parking_ali'),
         ('__setup__.stock_putaway_strat_parking_materiel',
          'specific_data.product_categ_materiel',
-         '__setup__.stock_location_parking_medoc'),
-        ('__setup__.stock_putaway_strat_parking_materiel',
-         'specific_data.product_categ_frigo',
-         '__setup__.stock_location_parking_frigo'),
+         '__setup__.stock_location_parking_materiel'),
         ]
     for xmlid, categ, loc in parking_strat:
         create_or_update(
@@ -286,9 +297,6 @@ def create_putaway(ctx):
         ('__setup__.stock_putaway_strat_onorder_medoc',
          'specific_data.product_categ_medoc',
          '__setup__.stock_location_order_medoc'),
-        ('__setup__.stock_putaway_strat_onorder_frigo',
-         'specific_data.product_categ_frigo',
-         '__setup__.stock_location_order_frigo'),
         ('__setup__.stock_putaway_strat_onorder_mat',
          'specific_data.product_categ_materiel',
          '__setup__.stock_location_order_mat'),
@@ -575,20 +583,24 @@ def create_routes(ctx):
     ref = ctx.env.ref
     types = [
         {'xmlid': '__setup__.stock_location_route_pick_materiel',
-         'name': 'Alcyon Belux SA: Pick (MAT)',
+         'name': 'Zone Matériel',
          'pull_ids': [(6, 0, ref('__setup__.procurement_rule_materiel').ids)],
+         'product_selectable': False,
          },
         {'xmlid': '__setup__.stock_location_route_pick_ali',
-         'name': 'Alcyon Belux SA: Pick (ALI)',
+         'name': 'Zone Aliments',
          'pull_ids': [(6, 0, ref('__setup__.procurement_rule_ali').ids)],
+         'product_selectable': False,
          },
         {'xmlid': '__setup__.stock_location_route_pick_medoc',
-         'name': 'Alcyon Belux SA: Pick (MED)',
+         'name': 'Zone Médicaments',
          'pull_ids': [(6, 0, ref('__setup__.procurement_rule_medoc').ids)],
+         'product_selectable': True,
          },
         {'xmlid': '__setup__.stock_location_route_pick_froid',
-         'name': 'Alcyon Belux SA: Pick (FROID)',
+         'name': 'Zone FROID / FRIGO',
          'pull_ids': [(6, 0, ref('__setup__.procurement_rule_froid').ids)],
+         'product_selectable': True,
          },
     ]
     for record in types:
@@ -596,7 +608,6 @@ def create_routes(ctx):
         record.update({
          'sequence': 20,
          'product_categ_selectable': True,
-         'product_selectable': False,
         })
         create_or_update(ctx, 'stock.location.route', xmlid, record)
 
@@ -611,8 +622,6 @@ def assign_route_categories(ctx):
                '__setup__.stock_location_route_pick_ali'),
               ('specific_data.product_categ_medoc',
                '__setup__.stock_location_route_pick_medoc'),
-              ('specific_data.product_categ_frigo',
-               '__setup__.stock_location_route_pick_froid'),
               ]
     for category_xmlid, route_xmlid in categs:
         ref(category_xmlid).route_ids = [(6, 0, ref(route_xmlid).ids)]
@@ -623,10 +632,10 @@ def main(ctx):
     """ Configuring logistics """
     company_settings(ctx)
     activate_options(ctx)
-    set_delivery_pick_ship(ctx)
+    warehouse_settings(ctx)
     create_locations(ctx)
-    create_putaway(ctx)
     create_picking_types(ctx)
     create_procurement_rules(ctx)
     create_routes(ctx)
+    create_putaway(ctx)
     assign_route_categories(ctx)